Move all virt-launcher live migration code into a separate files
