+++ pkg/virt-launcher/virtwrap/BUILD.bazel
@@ -4,8 +4,9 @@ go_library(
     name = "go_default_library",
     srcs = [
         "generated_mock_manager.go",
+        "live-migration-source.go",
+        "live-migration-target.go",
         "manager.go",
-        "postcopy.go",
     ],
     importpath = "kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap",
     visibility = ["//visibility:public"],
